Update index.md
There is no need to export the bot token and API token if we are using the cloud version of the Botkube, instead we just need to copy and paste them right into the webpage, so i added an instruction about that. More importantly i thought that people might get confused with the cloud version and the manual installation instructions.